+++ crates/storage/store_db/mod.rs
@@ -1,3 +1,5 @@
 pub mod in_memory;
 #[cfg(feature = "libmdbx")]
 pub mod libmdbx;
+#[cfg(feature = "rocksdb")]
+pub mod rocksdb;