+++ crates/storage/trie_db/mod.rs
@@ -6,6 +6,10 @@ pub mod libmdbx_dupsort;
 pub mod libmdbx_dupsort_locked;
 #[cfg(feature = "libmdbx")]
 pub mod libmdbx_locked;
+#[cfg(feature = "rocksdb")]
+pub mod rocksdb;
+#[cfg(feature = "rocksdb")]
+pub mod rocksdb_locked;
 #[cfg(test)]
 mod test_utils;
 pub mod utils;